decode-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
numpy==1.26.0
scipy==1.11.3
pyais==2.6.0
Pillow==10.1.0
paho-mqtt==1.6.1
python-multipart==0.0.6
